eTags working
Worked under the assumption that we are implementing eTags because the request otherwise returns a lot of data and makes a big database call.

So I chose to etag properties of the dataset which could be interrogated quickly without reading all the dataset. These were the max on an indexed created_datetime column and count on the dataset. These should be near instantaneous because the date column is indexed and many database engines cache the count of records where the right indexes are in play.